requirements.txt
dill==0.3.8
joblib==1.2.0
libnum==1.7.1
matplotlib==3.7.5
numpy==1.24.4
pandas==1.3.5
pycryptodome==3.20.0
pycryptodomex==3.17
PyNaCl==1.5.0
PyYAML==6.0.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.9.0
six==1.12.0
sympy==1.12
torch==1.12.1+cu113
torchvision==0.13.1
utilitybelt==0.2.6
